Simplify bench code
